Renamed domain to installer
