+++ src/xmipp3_installer/installer/handlers/shell_handler.py
